kube-proxy: ensure minimum kernel version for nftables mode
The nftables implementation makes use of concatenation of ranges when
creating the set "firewall-allow", but the support was not available
before kernel 5.6. Running nftables mode with older kernels doesn't work
and the failure log isn't very clear.

The commit checks the kernel version and returns a clear error if it
doesn't meet the requirement.

Signed-off-by: Quan Tian <qtian@vmware.com>