kube-proxy: fix LoadBalancerSourceRanges not working for nftables mode
Prerviously, the firewall-check chain was run in input, forward, and
output hook but not prerouting hook. When the LoadBalancer traffic
arrived at forward hook, it had been DNATed to endpoint IP and port, so
the firewall-check chain didn't take effect, traffic from out of
LoadBalancerSourceRanges was not dropped.

Another problem is that the firewall-allow set used by the
firewall-check chain makes use of concatenation of ranges, which was not
available before kernel 5.6.

The commit removes the firewall-check chain and related sets, and
leverages the service-ips map's verdict to jump to the service's
separate firewall chain, where the packet will be checked and dropped if
the source is not in the LoadBalancerSourceRanges.

It should improve the overall performance because it removes two service
specific sets and rules querying them.

Due to the removal of the firewall-allow set, nftables mode can run with
earlier kernels.

Signed-off-by: Quan Tian <qtian@vmware.com>